+++ src/page-ops.c
@@ -732,17 +732,17 @@ __hot int __must_check_result page_dirty(MDBX_txn *txn, page_t *mp, size_t npage
 
 void recalculate_subpage_thresholds(MDBX_env *env) {
   size_t whole = env->leaf_nodemax - NODESIZE;
-  env->subpage_limit = (whole * env->options.subpage.limit + 32767) >> 16;
+  env->subpage_limit = (uint16_t)((whole * env->options.subpage.limit + 32767) >> 16);
   whole = env->subpage_limit;
-  env->subpage_reserve_limit = (whole * env->options.subpage.reserve_limit + 32767) >> 16;
+  env->subpage_reserve_limit = (uint16_t)((whole * env->options.subpage.reserve_limit + 32767) >> 16);
   eASSERT0(env, env->leaf_nodemax >= env->subpage_limit + NODESIZE);
   eASSERT0(env, env->subpage_limit >= env->subpage_reserve_limit);
 
   whole = env->leaf_nodemax;
-  env->subpage_room_threshold = (whole * env->options.subpage.room_threshold + 32767) >> 16;
-  env->subpage_reserve_prereq = (whole * env->options.subpage.reserve_prereq + 32767) >> 16;
+  env->subpage_room_threshold = (uint16_t)((whole * env->options.subpage.room_threshold + 32767) >> 16);
+  env->subpage_reserve_prereq = (uint16_t)((whole * env->options.subpage.reserve_prereq + 32767) >> 16);
   if (env->subpage_room_threshold + env->subpage_reserve_limit > (intptr_t)page_space(env))
-    env->subpage_reserve_prereq = page_space(env);
+    env->subpage_reserve_prereq = (uint16_t)page_space(env);
   else if (env->subpage_reserve_prereq < env->subpage_room_threshold + env->subpage_reserve_limit)
     env->subpage_reserve_prereq = env->subpage_room_threshold + env->subpage_reserve_limit;
   eASSERT0(env, env->subpage_reserve_prereq >= env->subpage_room_threshold + env->subpage_reserve_limit);
